Moodify Final/Moodify/moodify/requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
tensorflow==2.6.0
keras==2.6.0
spotipy==2.19.0
python-dotenv==0.19.0